Fix integration test
